split out arge helper funcs
